fix(opencode-notifier-apprise): use plain JS instead of TypeScript bundle
OpenCode's plugin loader had issues with the bun-bundled output.
Switching to plain JS that OpenCode loads directly:

- Replace TypeScript source with plugin.js
- Remove package.json, tsconfig.json, src/ (no build step needed)
- Simplify home-manager module to just copy plugin.js
- OpenCode auto-discovers ~/.config/opencode/plugin/*.js
